Add user management module

+++ web/app.py
@@ -1,7 +1,24 @@
 from flask import Flask, render_template, request, redirect, session
-from app.database_sqlite import init_db, add_student, get_students, delete_student, update_student
-from web.auth import login_user, is_logged_in, logout_user
-
+from app.database_sqlite import (
+    init_db,
+    add_student,
+    get_students,
+    delete_student,
+    update_student,
+    get_users,
+    add_user,
+    delete_user,
+    update_user_role,
+    total_students,
+    total_users,
+    recent_students
+)
+from web.auth import (
+    login_user,
+    is_logged_in,
+    is_admin,
+    logout_user
+)
 app = Flask(__name__)
 app.secret_key = "mysecretkey"
 
@@ -37,7 +54,26 @@ def home():
         return redirect("/login")
 
     students = get_students()
-    return render_template("index.html", students=students, total=len(students), query="")
+    student_count = total_students()
+    user_count = total_users()
+    recent = recent_students()
+    return render_template(
+
+    "index.html",
+
+    students=students,
+
+    total=len(students),
+
+    query="",
+
+    student_count=student_count,
+
+    user_count=user_count,
+
+    recent=recent
+
+    )
 
 @app.route("/add", methods=["POST"])
 def add():
@@ -52,8 +88,12 @@ def add():
 
 @app.route("/delete/<int:roll>")
 def delete(roll):
+
     if not is_logged_in():
         return redirect("/login")
+
+    if not is_admin():
+        return "⛔ Access Denied", 403
 
     delete_student(roll)
     return redirect("/")
@@ -85,6 +125,57 @@ def search():
     ]
 
     return render_template("index.html", students=filtered, total=len(students), query=query)
+
+@app.route("/users")
+def users():
+
+    if not is_logged_in():
+        return redirect("/login")
+
+    if not is_admin():
+        return "Access Denied", 403
+
+    users = get_users()
+
+    return render_template("users.html", users=users)
+
+@app.route("/add_user", methods=["POST"])
+def add_new_user():
+
+    if not is_logged_in():
+        return redirect("/login")
+
+    if not is_admin():
+        return "Access Denied", 403
+
+    username = request.form["username"]
+    password = request.form["password"]
+    role = request.form["role"]
+
+    add_user(username, password)
+
+    # Role update
+    users = get_users()
+
+    for user in users:
+        if user[1] == username:
+            update_user_role(user[0], role)
+            break
+
+    return redirect("/users")
+
+@app.route("/delete_user/<int:user_id>")
+def remove_user(user_id):
+
+    if not is_logged_in():
+        return redirect("/login")
+
+    if not is_admin():
+        return "Access Denied", 403
+
+    delete_user(user_id)
+
+    return redirect("/users")
 
 if __name__ == "__main__":
     app.run(debug=True)
